Tidy up WASM renderer code
This tries to break up the massive game-of-life.go into smaller files,
and generally tidy up the renderer code for the rubik's cube

+++ wasm-demo/games/life/model/context.go
@@ -19,7 +19,7 @@ type LifeContext struct {
 	Zoom           float32
 	Paused         bool
 	SimulationSize int
-	Tps            int
+	Tps            float32
 	NeedsRestart   bool // will probably remove this one
 	Loop           bool
